Update README.md
ForesTech is a project using a soulbound token to tokenise carbon emissions. This wil be achieved by planned tree planting.  An ERC20 token will be issued to individual and investors/sponsors. Afterwards, a governance DAO will nominate planned areas of planting and how much tokens will be issued to the project.